OSPF (Open Shortest Path First)
Consider a data communication network that must route data packets
(email, MP3 les, or video
les, for example). Such a network consists of routers connected by
physical cables or links. A
router can act as a source, a destination, or a forwarder of data
packets. We can model a network
as a graph with each router corresponding to a vertex and the link or
physical connection between
two routers corresponding to a pair of directed edges between the
vertices.
A network that follows the OSPF (Open Shortest Path First) protocol
routes packets using
Dijkstra's shortest path algorithm. The criteria used to compute the
weight corresponding to a
link can include the time taken for data transmission, reliability of
the link, transmission cost, and
available bandwidth. Typically each router has a complete representation
of the network graph
and associated information available to it.
For the purposes of this project, each link has associated with it the
transmission time taken
for data to get from the vertex at one end to the vertex at the other
end. You will compute the
best path using the criterion of minimizing the total time taken for
data to reach the destination.
The shortest time path minimizes the sum of the transmission times of
the links along the path.
The network topology can change dynamically based on the state of the
links and the routers.
For example, a link may go down when the corresponding cable is cut, and
a vertex may go down
when the corresponding router crashes. In addition to these transient
changes, changes to a network
occur when a link is added or removed.